requirements.txt
pandas
openpyxl
streamlit
google-generativeai
pymongo
python-dotenv